feat: enhance application and deployment management APIs
- Added new API endpoints for creating, updating, retrieving applications and deployments.
- Implemented corresponding handler functions for each endpoint.
- Introduced logic layers for handling application and deployment operations.
- Updated routing to integrate new handlers and improve API structure.
