+++ backend/internal/clients/diagnosis/client.go
@@ -1,0 +1,1 @@
+package diagnosis
