+++ backend/internal/handler/apps/createapphandler.go
@@ -1,27 +1,27 @@
-package handler
+package apps
 
 import (
 	"net/http"
 
 	"github.com/Z3Labs/Hackathon/backend/common/errorx"
 	"github.com/Z3Labs/Hackathon/backend/common/httpresp"
-	"github.com/Z3Labs/Hackathon/backend/internal/logic"
+	"github.com/Z3Labs/Hackathon/backend/internal/logic/apps"
 	"github.com/Z3Labs/Hackathon/backend/internal/svc"
 	"github.com/Z3Labs/Hackathon/backend/internal/types"
 
 	"github.com/zeromicro/go-zero/rest/httpx"
 )
 
-func GetDeploymentDetailHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
+func CreateAppHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		var req types.GetDeploymentDetailReq
+		var req types.CreateAppReq
 		if err := httpx.Parse(r, &req); err != nil {
 			httpresp.HttpErr(w, r, errorx.NewStatCodeError(http.StatusBadRequest, 2, err.Error()))
 			return
 		}
 
-		l := logic.NewGetDeploymentDetailLogic(r.Context(), svcCtx)
-		resp, err := l.GetDeploymentDetail(&req)
+		l := apps.NewCreateAppLogic(r.Context(), svcCtx)
+		resp, err := l.CreateApp(&req)
 
 		httpresp.Http(w, r, resp, err)
