+++ backend/internal/handler/apps/getapplisthandler.go
@@ -1,11 +1,11 @@
-package handler
+package apps
 
 import (
 	"net/http"
 
 	"github.com/Z3Labs/Hackathon/backend/common/errorx"
 	"github.com/Z3Labs/Hackathon/backend/common/httpresp"
-	"github.com/Z3Labs/Hackathon/backend/internal/logic"
+	"github.com/Z3Labs/Hackathon/backend/internal/logic/apps"
 	"github.com/Z3Labs/Hackathon/backend/internal/svc"
 	"github.com/Z3Labs/Hackathon/backend/internal/types"
 
@@ -20,7 +20,7 @@ func GetAppListHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
 			return
 		}
 
-		l := logic.NewGetAppListLogic(r.Context(), svcCtx)
+		l := apps.NewGetAppListLogic(r.Context(), svcCtx)
 		resp, err := l.GetAppList(&req)
 
 		httpresp.Http(w, r, resp, err)
